requirements.txt
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Werkzeug==1.0.1
pip~=20.2.3
lxml~=4.6.2
Jinja2~=2.11.2
itsdangerous~=1.1.0
click~=7.1.2
python-dotenv~=0.17.1
MarkupSafe~=1.1.1
visitor~=0.1.3
certifi~=2020.6.20
chardet~=3.0.4
urllib3~=1.25.10
six~=1.15.0
dominate~=2.5.2
idna~=2.10
setuptools~=49.2.1
WTForms~=2.3.3
requests~=2.24.0